requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.8
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==10.0.0
greencache==1.0.0
htmq==1.0.0
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
redcmd==1.2.10
redlib==1.5.7
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
vote==1.0.2
zipp==3.7.0
zope.interface==5.4.0
